+++ pipeline/config/targets/build-shapes-ksj.ts
@@ -12,4 +12,5 @@ export default [
   'yurikamome',
   'mir-train',
   'tama-monorail',
+  'twr-rinkai',
 ];